+++ bioptim/optimization/receding_horizon_optimization.py
@@ -6,7 +6,7 @@ from time import perf_counter
 import numpy as np
 
 from .optimal_control_program import OptimalControlProgram
-from bioptim.optimization.solution.solution import Solution
+from ..optimization.solution.solution import Solution
 from ..dynamics.configure_problem import Dynamics, DynamicsList
 from ..limits.constraints import ConstraintFcn, ConstraintList
 from ..limits.objective_functions import ObjectiveFcn, ObjectiveList
@@ -14,7 +14,7 @@ from ..limits.path_conditions import InitialGuessList, BoundsList
 from ..misc.enums import SolverType, InterpolationType, MultiCyclicCycleSolutions, ControlType
 from ..optimization.parameters import ParameterList
 from ..interfaces.solver_options import Solver
-from ..interfaces.biomodel import BioModel
+from ..models.protocols.biomodel import BioModel
 
 
 class RecedingHorizonOptimization(OptimalControlProgram):
